+++ components/datetime/src/scaffold/calendar.rs
@@ -49,7 +49,7 @@ impl CldrCalendar for Buddhist {
     type SkeletaV1 = DatetimePatternsDateBuddhistV1;
 }
 
-impl CldrCalendar for Chinese {
+impl CldrCalendar for LunarChinese<chinese::China> {
     type YearNamesV1 = DatetimeNamesYearChineseV1;
     type MonthNamesV1 = DatetimeNamesMonthChineseV1;
     type SkeletaV1 = DatetimePatternsDateChineseV1;
@@ -61,7 +61,7 @@ impl CldrCalendar for Coptic {
     type SkeletaV1 = DatetimePatternsDateCopticV1;
 }
 
-impl CldrCalendar for Dangi {
+impl CldrCalendar for LunarChinese<chinese::Dangi> {
     type YearNamesV1 = DatetimeNamesYearDangiV1;
     type MonthNamesV1 = DatetimeNamesMonthDangiV1;
     type SkeletaV1 = DatetimePatternsDateDangiV1;
@@ -123,9 +123,9 @@ impl CldrCalendar for Roc {
 
 impl UnstableSealed for () {}
 impl UnstableSealed for Buddhist {}
-impl UnstableSealed for Chinese {}
+impl UnstableSealed for LunarChinese<chinese::China> {}
 impl UnstableSealed for Coptic {}
-impl UnstableSealed for Dangi {}
+impl UnstableSealed for LunarChinese<chinese::Dangi> {}
 impl UnstableSealed for Ethiopian {}
 impl UnstableSealed for Gregorian {}
 impl UnstableSealed for Hebrew {}
@@ -216,9 +216,9 @@ pub trait IntoFormattableAnyCalendar: CldrCalendar + IntoAnyCalendar {}
 
 // keep in sync with FormattableAnyCalendarKind
 impl IntoFormattableAnyCalendar for Buddhist {}
-impl IntoFormattableAnyCalendar for Chinese {}
+impl IntoFormattableAnyCalendar for LunarChinese<chinese::China> {}
 impl IntoFormattableAnyCalendar for Coptic {}
-impl IntoFormattableAnyCalendar for Dangi {}
+impl IntoFormattableAnyCalendar for LunarChinese<chinese::Dangi> {}
 impl IntoFormattableAnyCalendar for Ethiopian {}
 impl IntoFormattableAnyCalendar for Gregorian {}
 impl IntoFormattableAnyCalendar for Hebrew {}
@@ -357,9 +357,9 @@ impl FormattableAnyCalendar {
         use FormattableAnyCalendarKind::*;
         let any_calendar = match kind {
             Buddhist => AnyCalendar::Buddhist(cal::Buddhist),
-            Chinese => AnyCalendar::Chinese(cal::Chinese::new()),
+            Chinese => AnyCalendar::Chinese(cal::LunarChinese::new_china()),
             Coptic => AnyCalendar::Coptic(cal::Coptic),
-            Dangi => AnyCalendar::Dangi(cal::Dangi::new()),
+            Dangi => AnyCalendar::Dangi(cal::LunarChinese::new_dangi()),
             Ethiopian => AnyCalendar::Ethiopian(cal::Ethiopian::new()),
             EthiopianAmeteAlem => AnyCalendar::Ethiopian(cal::Ethiopian::new_with_era_style(
                 cal::EthiopianEraStyle::AmeteAlem,
@@ -394,9 +394,9 @@ impl FormattableAnyCalendar {
         use FormattableAnyCalendarKind::*;
         let any_calendar = match kind {
             Buddhist => AnyCalendar::Buddhist(cal::Buddhist),
-            Chinese => AnyCalendar::Chinese(cal::Chinese::new()),
+            Chinese => AnyCalendar::Chinese(cal::LunarChinese::new_china()),
             Coptic => AnyCalendar::Coptic(cal::Coptic),
-            Dangi => AnyCalendar::Dangi(cal::Dangi::new()),
+            Dangi => AnyCalendar::Dangi(cal::LunarChinese::new_dangi()),
             Ethiopian => AnyCalendar::Ethiopian(cal::Ethiopian::new()),
             EthiopianAmeteAlem => AnyCalendar::Ethiopian(cal::Ethiopian::new_with_era_style(
                 cal::EthiopianEraStyle::AmeteAlem,
@@ -432,9 +432,9 @@ impl FormattableAnyCalendar {
         use FormattableAnyCalendarKind::*;
         let any_calendar = match kind {
             Buddhist => AnyCalendar::Buddhist(cal::Buddhist),
-            Chinese => AnyCalendar::Chinese(cal::Chinese::new()),
+            Chinese => AnyCalendar::Chinese(cal::LunarChinese::new_china()),
             Coptic => AnyCalendar::Coptic(cal::Coptic),
-            Dangi => AnyCalendar::Dangi(cal::Dangi::new()),
+            Dangi => AnyCalendar::Dangi(cal::LunarChinese::new_dangi()),
             Ethiopian => AnyCalendar::Ethiopian(cal::Ethiopian::new()),
             EthiopianAmeteAlem => AnyCalendar::Ethiopian(cal::Ethiopian::new_with_era_style(
                 cal::EthiopianEraStyle::AmeteAlem,
@@ -572,9 +572,9 @@ where
 
 impl CalMarkers<YearNamesV1> for FullDataCalMarkers {
     type Buddhist = <Buddhist as CldrCalendar>::YearNamesV1;
-    type Chinese = <Chinese as CldrCalendar>::YearNamesV1;
+    type Chinese = <LunarChinese<chinese::China> as CldrCalendar>::YearNamesV1;
     type Coptic = <Coptic as CldrCalendar>::YearNamesV1;
-    type Dangi = <Dangi as CldrCalendar>::YearNamesV1;
+    type Dangi = <LunarChinese<chinese::Dangi> as CldrCalendar>::YearNamesV1;
     type Ethiopian = <Ethiopian as CldrCalendar>::YearNamesV1;
     type Gregorian = <Gregorian as CldrCalendar>::YearNamesV1;
     type Hebrew = <Hebrew as CldrCalendar>::YearNamesV1;
@@ -587,9 +587,9 @@ impl CalMarkers<YearNamesV1> for FullDataCalMarkers {
 
 impl CalMarkers<MonthNamesV1> for FullDataCalMarkers {
     type Buddhist = <Buddhist as CldrCalendar>::MonthNamesV1;
-    type Chinese = <Chinese as CldrCalendar>::MonthNamesV1;
+    type Chinese = <LunarChinese<chinese::China> as CldrCalendar>::MonthNamesV1;
     type Coptic = <Coptic as CldrCalendar>::MonthNamesV1;
-    type Dangi = <Dangi as CldrCalendar>::MonthNamesV1;
+    type Dangi = <LunarChinese<chinese::Dangi> as CldrCalendar>::MonthNamesV1;
     type Ethiopian = <Ethiopian as CldrCalendar>::MonthNamesV1;
     type Gregorian = <Gregorian as CldrCalendar>::MonthNamesV1;
     type Hebrew = <Hebrew as CldrCalendar>::MonthNamesV1;
@@ -602,9 +602,9 @@ impl CalMarkers<MonthNamesV1> for FullDataCalMarkers {
 
 impl CalMarkers<ErasedPackedPatterns> for FullDataCalMarkers {
     type Buddhist = <Buddhist as CldrCalendar>::SkeletaV1;
-    type Chinese = <Chinese as CldrCalendar>::SkeletaV1;
+    type Chinese = <LunarChinese<chinese::China> as CldrCalendar>::SkeletaV1;
     type Coptic = <Coptic as CldrCalendar>::SkeletaV1;
-    type Dangi = <Dangi as CldrCalendar>::SkeletaV1;
+    type Dangi = <LunarChinese<chinese::Dangi> as CldrCalendar>::SkeletaV1;
     type Ethiopian = <Ethiopian as CldrCalendar>::SkeletaV1;
     type Gregorian = <Gregorian as CldrCalendar>::SkeletaV1;
     type Hebrew = <Hebrew as CldrCalendar>::SkeletaV1;
